Move `VmSpace` tests to another file
